move useRapidAccessParam in hooks folder
