+++ packages/pn-personafisica-webapp/src/hooks/useRapidAccessParam.ts
@@ -1,0 +1,10 @@
+import { useMemo } from 'react';
+import { useSearchParams } from 'react-router-dom';
+
+import { getRapidAccessParam } from '@pagopa-pn/pn-commons/src/utility/routes.utility';
+
+export function useRapidAccessParam() {
+  const [params] = useSearchParams();
+
+  return useMemo(() => getRapidAccessParam(params), [params]);
+}
