+++ packages/pn-personafisica-webapp/src/navigation/RouteGuard.tsx
@@ -2,9 +2,10 @@ import { Outlet, useNavigate } from 'react-router-dom';
 
 import { AccessDenied } from '@pagopa-pn/pn-commons';
 
+import { useRapidAccessParam } from '../hooks/useRapidAccessParam';
 import { useAppSelector } from '../redux/hooks';
 import { RootState } from '../redux/store';
-import { goToLoginPortal, useRapidAccessParam } from './navigation.utility';
+import { goToLoginPortal } from './navigation.utility';
 import * as routes from './routes.const';
 
 const RouteGuard = () => {
